+++ lib/git/pkgs/commands/init.rb
@@ -125,6 +125,7 @@ module Git
                     purl: s[:purl],
                     requirement: s[:requirement],
                     dependency_type: s[:dependency_type],
+                    integrity: s[:integrity],
                     created_at: now,
                     updated_at: now
                   }
@@ -207,7 +208,8 @@ module Git
                     ecosystem: dep_info[:ecosystem],
                     purl: dep_info[:purl],
                     requirement: dep_info[:requirement],
-                    dependency_type: dep_info[:dependency_type]
+                    dependency_type: dep_info[:dependency_type],
+                    integrity: dep_info[:integrity]
                   }
                 end
                 snapshots_stored += snapshot.size
@@ -228,7 +230,8 @@ module Git
                   ecosystem: dep_info[:ecosystem],
                   purl: dep_info[:purl],
                   requirement: dep_info[:requirement],
-                  dependency_type: dep_info[:dependency_type]
+                  dependency_type: dep_info[:dependency_type],
+                  integrity: dep_info[:integrity]
                 }
               end
               snapshots_stored += snapshot.size